backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.1.2
pydantic==2.5.2
email-validator==2.1.0
pymongo==4.9.1
nltk==3.8.1
transformers==4.45.1
torch==2.4.1
beautifulsoup4==4.12.2
redis==5.0.1
aioredis==2.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.12.0
fakeredis==2.20.1
pytest-cov==4.1.0 